Fix crate name for cli
